+++ config/routes.rb
@@ -1,12 +1,17 @@
 Randomization::Application.routes.draw do
 
   resources :projects do
+    member do
+      post :generate_lists
+    end
+
     collection do
       post :add_treatment_arm
       post :add_stratification_factor
       post :add_option
-      post :generate_lists
     end
+
+    resources :assignments
   end
 
   devise_for :users, controllers: { registrations: 'contour/registrations', sessions: 'contour/sessions', passwords: 'contour/passwords', confirmations: 'contour/confirmations', unlocks: 'contour/unlocks' }, path_names: { sign_up: 'register', sign_in: 'login' }