+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - Randomization lists are now popuplated by shuffled blocks
   - Different block sizes and allocations can now be specified
   - Randomization goal can now be specified
+  - Randomization lists are now saved once generated, and can be regenerated
+    - Lists show up as individual project assignments
 
 ## 0.0.0 (August 21, 2013)